requirements.txt
Django==5.2.1
django-environ==0.12.0
django-redis==5.4.0
psycopg2-binary==2.9.10
redis==6.0.0
requests==2.32.3
pillow==11.2.1
gunicorn==23.0.0
djangorestframework==3.15.2
django-filter==24.2
celery==5.4.0
django-celery-results==2.6.0
pytest==8.3.2
pytest-django==4.8.0
amqp==5.2.0
kombu==5.4.0
pytest-cov==5.0.0